requirements.txt
fastmcp>=0.3.12
fast-agent-mcp>=0.3.12
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
pytz>=2023.3